Laboratory Manual added
Added what’s really a beta version of the lab manual. Not all of the
specifics from UW Bothell’s CSS 457 course have been removed. There are
still a lot of things I’d like to do with this, add to it, etc. So, it
is what it is for now.