Basic Matlab analog signal implementation
At this point, we have an analog signal class with basic functionality,
a sample and hold method to convert the analog signal to a discrete
signal, a quantizer function to convert a discrete signal to a digital
signal, and a plot function to make stair step plots of
discrete/digital signals. Next step is to start work on a Matlab
version of the lab manual.